updates product routes to include user auth

+++ backend/src/controllers/products.ts
@@ -1,11 +1,13 @@
 import { Request, Response } from "express";
 import ProductModel from "src/models/product";
+import UserModel from "src/models/user";
+import { AuthenticatedRequest, authenticateUser } from "src/validators/authUserMiddleware";
 import mongoose = require("mongoose");
 
 /**
  * get all the products in database
  */
-export const getProducts = async (req: Request, res: Response) => {
+export const getProducts = async (req: AuthenticatedRequest, res: Response) => {
   try {
     const products = await ProductModel.find();
     res.status(200).json(products);
@@ -16,7 +18,7 @@ export const getProducts = async (req: Request, res: Response) => {
 /**
  * get individual product thru product id
  */
-export const getProductById = async (req: Request, res: Response) => {
+export const getProductById = async (req: AuthenticatedRequest, res: Response) => {
   try {
     const id = req.params.id;
     if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -34,10 +36,11 @@ export const getProductById = async (req: Request, res: Response) => {
 /**
  * add product to database thru name, price, description, and userEmail
  */
-export const addProduct = async (req: Request, res: Response) => {
+export const addProduct = async (req: AuthenticatedRequest, res: Response) => {
   try {
-    const { name, price, description, userEmail } = req.body;
-
+    const { name, price, description } = req.body;
+    const userId = req.user.id;
+    const userEmail = req.user.userEmail;
     if (!name || !price || !userEmail) {
       return res.status(400).json({ message: "Name, price, and userEmail are required." });
     }
@@ -52,6 +55,9 @@ export const addProduct = async (req: Request, res: Response) => {
     });
 
     const savedProduct = await newProduct.save();
+    await UserModel.findByIdAndUpdate(userId, {
+      $push: { productList: savedProduct._id },
+    });
     res.status(201).json(savedProduct);
   } catch (error) {
     res.status(500).json({ message: "Error adding product", error });
@@ -60,19 +66,26 @@ export const addProduct = async (req: Request, res: Response) => {
 /**
  * delete product from database thru id
  */
-export const deleteProductById = async (req: Request, res: Response) => {
+export const deleteProductById = async (req: AuthenticatedRequest, res: Response) => {
   try {
     const id = req.params.id;
-
     if (!mongoose.Types.ObjectId.isValid(id)) {
       return res.status(400).json({ message: "Invalid ID format" });
+    }
+    const userId = req.user.id;
+    const user = await UserModel.findById(userId);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+    if (!user.productList.includes(id)) {
+      return res.status(400).json({ message: "User does not own this product" });
     }
 
     const deletedProduct = await ProductModel.findByIdAndDelete(id);
     if (!deletedProduct) {
       return res.status(404).json({ message: "Product not found" });
     }
-
+    await UserModel.findByIdAndUpdate(userId, { $pull: { productList: id } });
     res.status(200).json({ message: "Product successfully deleted", deletedProduct });
   } catch (error) {
     res.status(500).json({ message: "Error deleting product", error });
@@ -81,13 +94,21 @@ export const deleteProductById = async (req: Request, res: Response) => {
 /**
  * patch product in database thru id and updated parameters in req
  */
-export const updateProductById = async (req: Request, res: Response) => {
+export const updateProductById = async (req: AuthenticatedRequest, res: Response) => {
   try {
-    const updates = req.body;
     const id = req.params.id;
-
     if (!mongoose.Types.ObjectId.isValid(id)) {
       return res.status(400).json({ message: "Invalid ID format" });
+    }
+    const updates = req.body;
+    const userId = req.user.id;
+    const user = await UserModel.findById(userId);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
+    if (!user.productList.includes(id)) {
+      return res.status(400).json({ message: "User does not own this product" });
     }
 
     const updatedProduct = await ProductModel.findByIdAndUpdate(
